template: generate torizon configurations automatically
Reproduce our approach on the NXP branch. Most configurations will be
very similar except for the SOM and Display choice. We should use a
template file to easily manage them.

We still commit the generated configurations to the repository. We don't
want users to have to regenerate them manually.

The AM62 7 inch config was updated with the new template system.

Change-Id: I6a0fcd9b8fdb8b20b4150307c6d75d8b62b8d92f