template: set specific Toradex BSP version for verdin-imx95
Official support and pre-built images for the verdin-imx95
were added with Toradex BSP 7.4.0. Therefore, add specific customization
for this machine to build with a 7.4.0 BSP.

Note that the linux-imx submodule version was kept in 7.3.0 to keep
compatibility with other machine configurations.
Changes in device-tree headers between 7.3.0 and 7.4.0 does not impact
the verdin-imx95 support.

Change-Id: I51ace3ccc825be112a4c2ee1472f9223b9fd0423